fix not-in-ci _without_ resorting to a separate package.json
instead we can use prepack hook
prepare seems to be more for "always transpile this" instead of "only do this when building the package". things like coffeescript=>js where there's no tooling support for ANYTHING means you ALWAYS need to transpile. for ts you want to keep it in-ts as much as possible, and YOU CAN with things like tsx and vitest etc.